AQI_TRACKER
Project Members 
Vaishnavi Bhalerao, 
Pallavi Tambe, 
Mayuresh Kalkar
, Sahil Jadhav,

Project Information Our project is based on AQI that is Air Quality Index Tracker which gives the informtion of the location or the place which you wish to search for .Its important to know what kind of air you breath and live on , as they have major health effects on our body. Not only it gives information about the AQI but also gives the statistics about the air and also the presipitation ,humidity , water contain of the air,etc . And last it gives the informtion about the weather you wish to search for.